Fix Digest security for jaspi and non-jaspi authentication: must pass a Digest instance to be compared against the password stored for the user, as Digest has an algorithm, its not enough to just compare the bytes of the Digest.
Changed the method signature of LoginCallbackImpl to use Credential rather than char[] for credentials because not all credentials are bytes (cf Digest).

Introduced CredentialValidationCallback so that Credential can be passed back to be validated, as the PasswordValidationCallback only deals with char[] type of credentials (ie simple passwords).

Made similar changes to JAAS to handle more than simple char[] passwords.